test(125658): Remove testes do app escola e inclui no apo eol_servico
