+++ sme_sigpae_api/eol_servico/__tests__/test_utils.py
@@ -105,3 +105,71 @@ class TestEOLServicoSGP(TestCase):
 
         self.assertIn("API EOL do SGP está com erro", str(context.exception))
         self.assertIn("Status: 503", str(context.exception))
+
+
+def test_redefine_senha_coresso(client_autenticado_da_escola, monkeypatch):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_altera_senha",
+        lambda p1: mocked_response({}, 200),
+    )
+    response = EOLServicoSGP.redefine_senha("123456", "adminadmin")
+    assert response == "OK"
+
+
+def test_redefine_senha_coresso_eol_exception(
+    client_autenticado_da_escola, monkeypatch
+):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_altera_senha",
+        lambda p1: mocked_response({"erro": "EOL fora do ar"}, 400),
+    )
+    with pytest.raises(EOLException):
+        EOLServicoSGP.redefine_senha("123456", "adminadmin")
+
+
+def test_redefine_email_coresso(client_autenticado_da_escola, monkeypatch):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_altera_email_coresso",
+        lambda p1: mocked_response({}, 200),
+    )
+    response = EOLServicoSGP.redefine_email("123456", "adminadmin")
+    assert response == "OK"
+
+
+def test_redefine_email_coresso_eol_exception(
+    client_autenticado_da_escola, monkeypatch
+):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_altera_email_coresso",
+        lambda p1: mocked_response({"erro": "EOL fora do ar"}, 400),
+    )
+    with pytest.raises(EOLException):
+        EOLServicoSGP.redefine_email("123456", "adminadmin")
+
+
+def test_criar_usuario_coresso(client_autenticado_da_escola, monkeypatch):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_criar_usuario_coresso",
+        lambda p1, p2: mocked_response({}, 200),
+    )
+    response = EOLServicoSGP.cria_usuario_core_sso(
+        "123456", "FULANO DA SILVA", "fulano@silva.com"
+    )
+    assert response == "OK"
+
+
+def test_criar_usuario_coresso_eol_exception(client_autenticado_da_escola, monkeypatch):
+    monkeypatch.setattr(
+        EOLServicoSGP,
+        "chamada_externa_criar_usuario_coresso",
+        lambda p1, p2: mocked_response({"erro": "EOL fora do ar"}, 400),
+    )
+    with pytest.raises(EOLException):
+        EOLServicoSGP.cria_usuario_core_sso(
+            "123456", "FULANO DA SILVA", "fulano@silva.com"
+        )
